试卷、试题开发

+++ src/main/java/org/eocencle/dasislcy/entity/ExamquestionEntity.java
@@ -20,13 +20,13 @@ public class ExamquestionEntity {
      * 顺序
      */
     @Column(name = "SORT")
-    private Byte sort;
+    private Integer sort;
 
     /**
      * 题型 1、选择题
      */
     @Column(name = "QUESTION_TYPE")
-    private Byte questionType;
+    private Integer questionType;
 
     /**
      * 试题ID
@@ -83,7 +83,7 @@ public class ExamquestionEntity {
      *
      * @return SORT - 顺序
      */
-    public Byte getSort() {
+    public Integer getSort() {
         return sort;
     }
 
@@ -92,7 +92,7 @@ public class ExamquestionEntity {
      *
      * @param sort 顺序
      */
-    public void setSort(Byte sort) {
+    public void setSort(Integer sort) {
         this.sort = sort;
     }
 
@@ -101,7 +101,7 @@ public class ExamquestionEntity {
      *
      * @return QUESTION_TYPE - 题型 1、选择题
      */
-    public Byte getQuestionType() {
+    public Integer getQuestionType() {
         return questionType;
     }
 
@@ -110,7 +110,7 @@ public class ExamquestionEntity {
      *
      * @param questionType 题型 1、选择题
      */
-    public void setQuestionType(Byte questionType) {
+    public void setQuestionType(Integer questionType) {
         this.questionType = questionType;
     }
